I'm dropping this because of no community interest.
Everyone that was playing on the Minecraft server has left and no one has complained about it being down during unexpected events.

I'm not bringing this back. I have spent substantial weeks on a production-grade Minecraft server that only lead to no one to play on it.

Do not ask me to bring back or host a Minecraft server unless I'm being paid. I have done all of this in my spare time and with no pay.

Signed-off-by: June <june@eridan.me>